# docs/swagger.yaml
basePath: /api
definitions:
...
  github_com_faridlan_sikon-api_internal_delivery_http_dto.BankAccountCreateRequest:
    properties:
      account_name:
        example: PT SIKOn Konveksi
        type: string
      account_number:
        example: "1234567890"
        type: string
      bank_name:
        example: BCA
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    required:
    - account_name
    - account_number
    - bank_name
    type: object
  github_com_faridlan_sikon-api_internal_delivery_http_dto.BankAccountResponse:
    properties:
      account_name:
        example: PT SIKOn Konveksi
        type: string
      account_number:
        example: "1234567890"
        type: string
      bank_name:
        example: BCA
        type: string
      created_at:
        example: "2023-10-01T15:00:00Z"
        type: string
      id:
        example: 777e4567-e89b-12d3-a456-426614174000
        type: string
      updated_at:
        example: "2023-10-01T15:00:00Z"
        type: string
      user:
        allOf:
        - $ref: '#/definitions/github_com_faridlan_sikon-api_internal_delivery_http_dto.UserResponse'
        description: Relasi ke User
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
  github_com_faridlan_sikon-api_internal_delivery_http_dto.BankAccountUpdateRequest:
    properties:
      account_name:
        example: Budi Haryanto
        type: string
      account_number:
        example: "0987654321"
        type: string
      bank_name:
        example: Mandiri
        type: string
    type: object
...
  github_com_faridlan_sikon-api_internal_delivery_http_dto.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      image_url:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      phone:
        type: string
      role:
        type: string
      sort_order:
        type: integer
      status_text:
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_faridlan_sikon-api_internal_domain.PaginationMeta:
    properties:
      current_page:
        type: integer
      limit:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
  github_com_faridlan_sikon-api_internal_utils.EmptyObj:
    type: object
  github_com_faridlan_sikon-api_internal_utils.ErrorResponse:
    properties:
      detail:
        type: string
      error:
        example: pesan error
        type: string
    type: object
  ? github_com_faridlan_sikon-api_internal_utils.PaginatedResponse-github_com_faridlan_sikon-api_internal_delivery_http_dto_BankAccountResponse
  : properties:
      data:
        items:
          $ref: '#/definitions/github_com_faridlan_sikon-api_internal_delivery_http_dto.BankAccountResponse'
        type: array
      message:
        type: string
      meta:
        $ref: '#/definitions/github_com_faridlan_sikon-api_internal_domain.PaginationMeta'
    type: object
...
  ? github_com_faridlan_sikon-api_internal_utils.SuccessResponse-array_github_com_faridlan_sikon-api_internal_delivery_http_dto_BankAccountResponse
  : properties:
      data:
        items:
          $ref: '#/definitions/github_com_faridlan_sikon-api_internal_delivery_http_dto.BankAccountResponse'
        type: array
      message:
        type: string
    type: object
...
  ? github_com_faridlan_sikon-api_internal_utils.SuccessResponse-github_com_faridlan_sikon-api_internal_delivery_http_dto_BankAccountResponse
  : properties:
      data:
        $ref: '#/definitions/github_com_faridlan_sikon-api_internal_delivery_http_dto.BankAccountResponse'
      message:
        type: string
    type: object
...
  github_com_faridlan_sikon-api_internal_utils.SuccessResponse-github_com_faridlan_sikon-api_internal_utils_EmptyObj:
    properties:
      data:
        $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.EmptyObj'
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Ini adalah dokumentasi API untuk MVP ERP SIKOn.
  title: SIKOn API (Sistem Integrasi Konveksi Online)
  version: "1.0"
paths:
...
  /bank-accounts:
    get:
      description: Mengambil daftar seluruh rekening bank
      parameters:
      - default: 1
        description: Nomor Halaman
        in: query
        name: page
        type: integer
      - default: 10
        description: Batas Data per Halaman
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.PaginatedResponse-github_com_faridlan_sikon-api_internal_delivery_http_dto_BankAccountResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List All Bank Accounts
      tags:
      - Bank Accounts
    post:
      consumes:
      - application/json
      description: Menambahkan rekening bank (Perusahaan / Sales)
      parameters:
      - description: Data Rekening Baru
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_faridlan_sikon-api_internal_delivery_http_dto.BankAccountCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.SuccessResponse-github_com_faridlan_sikon-api_internal_delivery_http_dto_BankAccountResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create Bank Account
      tags:
      - Bank Accounts
  /bank-accounts/{id}:
    delete:
      description: Menghapus rekening secara permanen
      parameters:
      - description: Bank Account ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.SuccessResponse-github_com_faridlan_sikon-api_internal_utils_EmptyObj'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete Bank Account
      tags:
      - Bank Accounts
    get:
      description: Mengambil detail rekening bank berdasarkan ID
      parameters:
      - description: Bank Account ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.SuccessResponse-github_com_faridlan_sikon-api_internal_delivery_http_dto_BankAccountResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Bank Account by ID
      tags:
      - Bank Accounts
    put:
      consumes:
      - application/json
      description: Memperbarui informasi rekening bank
      parameters:
      - description: Bank Account ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Data Rekening yang Diperbarui
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_faridlan_sikon-api_internal_delivery_http_dto.BankAccountUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.SuccessResponse-github_com_faridlan_sikon-api_internal_delivery_http_dto_BankAccountResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update Bank Account
      tags:
      - Bank Accounts
  /bank-accounts/global:
    get:
      description: Mengambil daftar rekening khusus perusahaan (User ID null)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.SuccessResponse-array_github_com_faridlan_sikon-api_internal_delivery_http_dto_BankAccountResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Global Bank Accounts
      tags:
      - Bank Accounts
  /bank-accounts/user/{user_id}:
    get:
      description: Mengambil daftar rekening bank milik user tertentu
      parameters:
      - description: User ID (UUID)
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.SuccessResponse-array_github_com_faridlan_sikon-api_internal_delivery_http_dto_BankAccountResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get User's Bank Accounts
      tags:
      - Bank Accounts
...
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token dengan format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"