# internal/delivery/http/bank_account_handler.go
package http

import (
	"github.com/gofiber/fiber/v2"

	"github.com/faridlan/sikon-api/internal/delivery/http/dto"
	"github.com/faridlan/sikon-api/internal/domain"
	"github.com/faridlan/sikon-api/internal/utils"
)

type BankAccountHandler interface {
	CreateAccount(c *fiber.Ctx) error
	GetByID(c *fiber.Ctx) error
	UpdateAccount(c *fiber.Ctx) error
	ListAccounts(c *fiber.Ctx) error
	GetGlobalAccounts(c *fiber.Ctx) error
	GetUserAccounts(c *fiber.Ctx) error
	DeleteAccount(c *fiber.Ctx) error
}

type bankAccountHandler struct {
	bankAccountUsecase domain.BankAccountUsecase
}

func NewBankAccountHandler(bu domain.BankAccountUsecase) BankAccountHandler {
	return &bankAccountHandler{
		bankAccountUsecase: bu,
	}
}

// @Summary Create Bank Account
// @Description Menambahkan rekening bank (Perusahaan / Sales)
// @Tags Bank Accounts
// @Accept json
// @Produce json
// @Param request body dto.BankAccountCreateRequest true "Data Rekening Baru"
// @Success 201 {object} utils.SuccessResponse[dto.BankAccountResponse]
// @Failure 400 {object} utils.ErrorResponse
// @Failure 500 {object} utils.ErrorResponse
// @Failure 401 {object} utils.ErrorResponse "Unauthorized"
// @Security BearerAuth
// @Router /bank-accounts [post]
func (h *bankAccountHandler) CreateAccount(c *fiber.Ctx) error {
	var req dto.BankAccountCreateRequest

	if err := c.BodyParser(&req); err != nil {
		return utils.SendError(c, fiber.StatusBadRequest, "Gagal memparsing request body")
	}

	if err := utils.ValidateStruct(&req); err != nil {
		return utils.SendError(c, fiber.StatusBadRequest, err.Error())
	}

	// Gunakan Input Struct yang baru dibuat di domain
	domainReq := domain.BankAccountCreateInput{
		UserID:        req.UserID,
		BankName:      req.BankName,
		AccountNumber: req.AccountNumber,
		AccountName:   req.AccountName,
	}

	account, err := h.bankAccountUsecase.CreateAccount(c.Context(), domainReq)
	if err != nil {
		return utils.HandleDomainError(c, err)
	}

	return utils.SendSuccess(c, fiber.StatusCreated, "Berhasil menambahkan rekening", dto.ToBankAccountResponse(account))
}

// @Summary Get Bank Account by ID
// @Description Mengambil detail rekening bank berdasarkan ID
// @Tags Bank Accounts
// @Produce json
// @Param id path string true "Bank Account ID (UUID)"
// @Success 200 {object} utils.SuccessResponse[dto.BankAccountResponse]
// @Failure 400 {object} utils.ErrorResponse
// @Failure 404 {object} utils.ErrorResponse
// @Failure 500 {object} utils.ErrorResponse
// @Failure 401 {object} utils.ErrorResponse "Unauthorized"
// @Security BearerAuth
// @Router /bank-accounts/{id} [get]
func (h *bankAccountHandler) GetByID(c *fiber.Ctx) error {
	id := c.Params("id")
	if err := utils.ValidateUUID(id, "id"); err != nil {
		return utils.SendError(c, fiber.StatusBadRequest, err.Error())
	}
	account, err := h.bankAccountUsecase.GetAccount(c.Context(), id)
	if err != nil {
		return utils.HandleDomainError(c, err)
	}
	return utils.SendSuccess(c, fiber.StatusOK, "Berhasil mengambil data rekening", dto.ToBankAccountResponse(account))
}

// @Summary Update Bank Account
// @Description Memperbarui informasi rekening bank
// @Tags Bank Accounts
// @Accept json
// @Produce json
// @Param id path string true "Bank Account ID (UUID)"
// @Param request body dto.BankAccountUpdateRequest true "Data Rekening yang Diperbarui"
// @Success 200 {object} utils.SuccessResponse[dto.BankAccountResponse]
// @Failure 400 {object} utils.ErrorResponse
// @Failure 500 {object} utils.ErrorResponse
// @Failure 401 {object} utils.ErrorResponse "Unauthorized"
// @Security BearerAuth
// @Router /bank-accounts/{id} [put]
func (h *bankAccountHandler) UpdateAccount(c *fiber.Ctx) error {
	id := c.Params("id")
	if err := utils.ValidateUUID(id, "id"); err != nil {
		return utils.SendError(c, fiber.StatusBadRequest, err.Error())
	}

	var req dto.BankAccountUpdateRequest
	if err := c.BodyParser(&req); err != nil {
		return utils.SendError(c, fiber.StatusBadRequest, "Gagal memparsing request body")
	}

	if err := utils.ValidateStruct(&req); err != nil {
		return utils.SendError(c, fiber.StatusBadRequest, err.Error())
	}

	// Gunakan Input Struct yang baru dibuat di domain
	domainReq := domain.BankAccountUpdateInput{
		BankName:      req.BankName,
		AccountNumber: req.AccountNumber,
		AccountName:   req.AccountName,
	}

	account, err := h.bankAccountUsecase.UpdateAccount(c.Context(), id, domainReq)
	if err != nil {
		return utils.HandleDomainError(c, err)
	}

	return utils.SendSuccess(c, fiber.StatusOK, "Berhasil memperbarui rekening", dto.ToBankAccountResponse(account))
}

// @Summary List All Bank Accounts
// @Description Mengambil daftar seluruh rekening bank
// @Tags Bank Accounts
// @Produce json
// @Param page query int false "Nomor Halaman" default(1)
// @Param limit query int false "Batas Data per Halaman" default(10)
// @Success 200 {object} utils.PaginatedResponse[dto.BankAccountResponse]
// @Failure 500 {object} utils.ErrorResponse
// @Failure 401 {object} utils.ErrorResponse "Unauthorized"
// @Security BearerAuth
// @Router /bank-accounts [get]
func (h *bankAccountHandler) ListAccounts(c *fiber.Ctx) error {
	page := c.QueryInt("page", 1)
	limit := c.QueryInt("limit", 10)

	query := domain.PaginationQuery{Page: page, Limit: limit}

	accounts, meta, err := h.bankAccountUsecase.ListAccounts(c.Context(), query)
	if err != nil {
		return utils.HandleDomainError(c, err)
	}

	return utils.SendSuccessPaginated(c, "Berhasil mengambil daftar rekening", dto.ToBankAccountResponseList(accounts), meta)
}

// @Summary Get Global Bank Accounts
// @Description Mengambil daftar rekening khusus perusahaan (User ID null)
// @Tags Bank Accounts
// @Produce json
// @Success 200 {object} utils.SuccessResponse[[]dto.BankAccountResponse]
// @Failure 500 {object} utils.ErrorResponse
// @Failure 401 {object} utils.ErrorResponse "Unauthorized"
// @Security BearerAuth
// @Router /bank-accounts/global [get]
func (h *bankAccountHandler) GetGlobalAccounts(c *fiber.Ctx) error {
	accounts, err := h.bankAccountUsecase.GetGlobalAccounts(c.Context())
	if err != nil {
		return utils.HandleDomainError(c, err)
	}

	return utils.SendSuccess(c, fiber.StatusOK, "Berhasil mengambil rekening global", dto.ToBankAccountResponseList(accounts))
}

// @Summary Get User's Bank Accounts
// @Description Mengambil daftar rekening bank milik user tertentu
// @Tags Bank Accounts
// @Produce json
// @Param user_id path string true "User ID (UUID)"
// @Success 200 {object} utils.SuccessResponse[[]dto.BankAccountResponse]
// @Failure 400 {object} utils.ErrorResponse
// @Failure 500 {object} utils.ErrorResponse
// @Failure 401 {object} utils.ErrorResponse "Unauthorized"
// @Security BearerAuth
// @Router /bank-accounts/user/{user_id} [get]
func (h *bankAccountHandler) GetUserAccounts(c *fiber.Ctx) error {
	userID := c.Params("user_id")
	if err := utils.ValidateUUID(userID, "user_id"); err != nil {
		return utils.SendError(c, fiber.StatusBadRequest, err.Error())
	}

	accounts, err := h.bankAccountUsecase.GetUserAccounts(c.Context(), userID)
	if err != nil {
		return utils.HandleDomainError(c, err)
	}

	return utils.SendSuccess(c, fiber.StatusOK, "Berhasil mengambil rekening user", dto.ToBankAccountResponseList(accounts))
}

// @Summary Delete Bank Account
// @Description Menghapus rekening secara permanen
// @Tags Bank Accounts
// @Produce json
// @Param id path string true "Bank Account ID (UUID)"
// @Success 200 {object} utils.SuccessResponse[utils.EmptyObj]
// @Failure 400 {object} utils.ErrorResponse
// @Failure 500 {object} utils.ErrorResponse
// @Failure 401 {object} utils.ErrorResponse "Unauthorized"
// @Security BearerAuth
// @Router /bank-accounts/{id} [delete]
func (h *bankAccountHandler) DeleteAccount(c *fiber.Ctx) error {
	id := c.Params("id")
	if err := utils.ValidateUUID(id, "id"); err != nil {
		return utils.SendError(c, fiber.StatusBadRequest, err.Error())
	}

	if err := h.bankAccountUsecase.DeleteAccount(c.Context(), id); err != nil {
		return utils.HandleDomainError(c, err)
	}

	return utils.SendSuccess(c, fiber.StatusOK, "Berhasil menghapus rekening", nil)
}